chore: bump to 0.6.0 + scattered clippy cleanups
Wire-format change in efd-proto (PROTO_VERSION byte) is a hard break
with any older deployed binary — bump minor to signal it and skip
over 0.5.0 (which was only ever in a reset-away commit but was
deployed to the Pi, so the version is burned). Also bump
dist/arch/PKGBUILD in lockstep.

Miscellaneous clippy cleanups picked up along the way:

  - efd-audio (alsa_out/usb_rx/usb_tx): pcm.recover now takes the
    errno directly; drop the `as i32` cast
  - efd-iq/drivers/fdm_duo: block_count % N == 0 → is_multiple_of
  - client/audio: same + drop dead volume() getter
  - client/ui/mod: drop dead set_zoom (all callers inlined the clamp
    logic themselves)
  - client/ui/controls: #[allow(dead_code)] on last_cmd — the field
    is cloned into per-control closures via Rc, so the struct's copy
    looks unread without the annotation

+++ crates/efd-audio/src/alsa_out.rs
@@ -104,7 +104,7 @@ fn run_alsa(
                 Ok(n) => offset += n,
                 Err(e) => {
                     warn!("ALSA write error: {e}, recovering");
-                    if let Err(e2) = pcm.recover(e.errno() as i32, true) {
+                    if let Err(e2) = pcm.recover(e.errno(), true) {
                         error!("ALSA recover failed: {e2}");
                         return Err(AudioError::Alsa(e));
                     }
