+++ crates/efd-audio/src/usb_rx.rs
@@ -86,7 +86,7 @@ fn run_usb_rx(
                 Ok(n) => offset += n,
                 Err(e) => {
                     warn!("USB RX read error: {e}, recovering");
-                    if let Err(e2) = pcm.recover(e.errno() as i32, true) {
+                    if let Err(e2) = pcm.recover(e.errno(), true) {
                         warn!("USB RX recover failed: {e2}");
                         return Err(AudioError::Alsa(e));
                     }
